feat(zcash): move types into zcash-shielded
