+++ .changeset/late-laws-behave.md
@@ -1,0 +1,6 @@
+---
+"@ledgerhq/zcash-shielded": minor
+"@ledgerhq/coin-bitcoin": minor
+---
+
+Move zcash types to zcash-shielded
