+++ libs/coin-modules/coin-bitcoin/package.json
@@ -54,6 +54,7 @@
     "@ledgerhq/logs": "workspace:^",
     "@ledgerhq/types-cryptoassets": "workspace:^",
     "@ledgerhq/types-live": "workspace:^",
+    "@ledgerhq/zcash-shielded": "workspace:^",
     "bchaddrjs": "^0.5.2",
     "bech32": "^1.1.3",
     "bignumber.js": "^9.1.2",
