+++ libs/coin-modules/zcash-shielded/package.json
@@ -67,7 +67,6 @@
   "dependencies": {
     "@ledgerhq/live-network": "workspace:^",
     "@ledgerhq/logs": "workspace:^",
-    "@ledgerhq/coin-bitcoin": "workspace:^",
     "@ledgerhq/zcash-decrypt": "0.1.0",
     "bignumber.js": "^9.1.2"
   }
